+++ lib/twilio-ruby/rest/api/v2010/account/call.rb
@@ -637,7 +637,7 @@ module Twilio
             end
 
             ##
-            # @return [feedback_summary.Status] The status
+            # @return [recording.Status] The status
             def status
               @properties['status']
             end