+++ lib/twilio-ruby/rest/api/v2010/account/call/recording.rb
@@ -248,6 +248,10 @@ module Twilio
                     'uri' => payload['uri'],
                     'encryption_details' => payload['encryption_details'],
                     'error_code' => payload['error_code'] == nil ? payload['error_code'] : payload['error_code'].to_i,
+                    'status' => payload['status'],
+                    'source' => payload['source'],
+                    'channels' => payload['channels'].to_i,
+                    'price_unit' => payload['price_unit'],
                 }
 
                 # Context
@@ -335,6 +339,30 @@ module Twilio
               # @return [String] The error_code
               def error_code
                 @properties['error_code']
+              end
+
+              ##
+              # @return [recording.Status] The status
+              def status
+                @properties['status']
+              end
+
+              ##
+              # @return [recording.Source] The source
+              def source
+                @properties['source']
+              end
+
+              ##
+              # @return [String] The channels
+              def channels
+                @properties['channels']
+              end
+
+              ##
+              # @return [String] The price_unit
+              def price_unit
+                @properties['price_unit']
               end
 
               ##