+++ lib/twilio-ruby/rest/preview/hosted_numbers/authorization_document.rb
@@ -27,6 +27,11 @@ module Twilio
           # Lists AuthorizationDocumentInstance records from the API as a list.
           # Unlike stream(), this operation is eager and will load `limit` records into
           # memory before returning.
+          # @param [String] email Email that this AuthorizationDocument will be sent to for
+          #   signing.
+          # @param [authorization_document.Status] status The Status of this
+          #   AuthorizationDocument. One of `opened`, `signing`, `signed`, `canceled`, or
+          #   `failed`.
           # @param [Integer] limit Upper limit for the number of records to return. stream()
           #    guarantees to never return more than limit.  Default is no limit
           # @param [Integer] page_size Number of records to fetch per request, when
@@ -34,14 +39,19 @@ module Twilio
           #    but a limit is defined, stream() will attempt to read the limit with the most
           #    efficient page size, i.e. min(limit, 1000)
           # @return [Array] Array of up to limit results
-          def list(limit: nil, page_size: nil)
-            self.stream(limit: limit, page_size: page_size).entries
+          def list(email: :unset, status: :unset, limit: nil, page_size: nil)
+            self.stream(email: email, status: status, limit: limit, page_size: page_size).entries
           end
 
           ##
           # Streams AuthorizationDocumentInstance records from the API as an Enumerable.
           # This operation lazily loads records as efficiently as possible until the limit
           # is reached.
+          # @param [String] email Email that this AuthorizationDocument will be sent to for
+          #   signing.
+          # @param [authorization_document.Status] status The Status of this
+          #   AuthorizationDocument. One of `opened`, `signing`, `signed`, `canceled`, or
+          #   `failed`.
           # @param [Integer] limit Upper limit for the number of records to return. stream()
           #    guarantees to never return more than limit. Default is no limit.
           # @param [Integer] page_size Number of records to fetch per request, when
@@ -49,10 +59,10 @@ module Twilio
           #    but a limit is defined, stream() will attempt to read the limit with the most
           #    efficient page size, i.e. min(limit, 1000)
           # @return [Enumerable] Enumerable that will yield up to limit results
-          def stream(limit: nil, page_size: nil)
+          def stream(email: :unset, status: :unset, limit: nil, page_size: nil)
             limits = @version.read_limits(limit, page_size)
 
-            page = self.page(page_size: limits[:page_size])
+            page = self.page(email: email, status: status, page_size: limits[:page_size])
 
             @version.stream(page, limit: limits[:limit], page_limit: limits[:page_limit])
           end
@@ -74,12 +84,19 @@ module Twilio
           ##
           # Retrieve a single page of AuthorizationDocumentInstance records from the API.
           # Request is executed immediately.
+          # @param [String] email Email that this AuthorizationDocument will be sent to for
+          #   signing.
+          # @param [authorization_document.Status] status The Status of this
+          #   AuthorizationDocument. One of `opened`, `signing`, `signed`, `canceled`, or
+          #   `failed`.
           # @param [String] page_token PageToken provided by the API
           # @param [Integer] page_number Page Number, this value is simply for client state
           # @param [Integer] page_size Number of records to return, defaults to 50
           # @return [Page] Page of AuthorizationDocumentInstance
-          def page(page_token: :unset, page_number: :unset, page_size: :unset)
+          def page(email: :unset, status: :unset, page_token: :unset, page_number: :unset, page_size: :unset)
             params = Twilio::Values.of({
+                'Email' => email,
+                'Status' => status,
                 'PageToken' => page_token,
                 'Page' => page_number,
                 'PageSize' => page_size,