+++ lib/twilio-ruby/rest/proxy/v1/service/session/interaction.rb
@@ -234,6 +234,13 @@ module Twilio
               end
 
               ##
+              # Deletes the InteractionInstance
+              # @return [Boolean] true if delete succeeds, true otherwise
+              def delete
+                @version.delete('delete', @uri)
+              end
+
+              ##
               # Provide a user friendly representation
               def to_s
                 context = @solution.map {|k, v| "#{k}: #{v}"}.join(',')
@@ -423,6 +430,13 @@ module Twilio
               # @return [InteractionInstance] Fetched InteractionInstance
               def fetch
                 context.fetch
+              end
+
+              ##
+              # Deletes the InteractionInstance
+              # @return [Boolean] true if delete succeeds, true otherwise
+              def delete
+                context.delete
               end
 
               ##